Renames and refactors PIM Azure role management functions
Renames the 'New-PIMAzureRoleAssignmentScheduleRequest' and 'Remove-PIMAzureRoleAssignmentScheduleRequest' functions to 'Start-PIMAzureEligibleRoleAssignment' and 'Stop-PIMAzureEligibleRoleAssignment' respectively.

Updates the function descriptions and examples to reflect the new naming and the actions of starting and stopping eligible role assignments.

This change improves clarity and aligns the function names with their intended purpose of activating and deactivating PIM Azure role assignments.