Refactors principal ID retrieval for PIM roles
Simplifies the retrieval of the principal ID in the `Start-PIMAzureEligibleRoleAssignment` and `Stop-PIMAzureEligibleRoleAssignment` functions.

It now prioritizes using the `servicePrincipalId` environment variable if it is set, avoiding reliance on parsing the current Azure context when a service principal is being used. This improves reliability and simplifies the logic for obtaining the principal ID.

Also, fixes validation script in `Stop-PIMAzureEligibleRoleAssignment` and adds a 5-minute check before role deactivation.